Rename AtAGlanceWidget to SmartAtAGlanceWidget with enhanced features
